remove codec
