--- api/v1beta2/codec.go
@@ -1,58 +1,0 @@
-package v1beta2
-
-import (
-	"encoding/json"
-	"fmt"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/yaml"
-)
-
-// ImageSpec defines OCI Image specifications.
-// +k8s:deepcopy-gen=true
-type ImageSpec struct {
-	// Repo defines the Image repo
-	Repo string `json:"repo,omitempty"`
-
-	// Name defines the Image name
-	Name string `json:"name,omitempty"`
-
-	// Ref is either a sha value, tag or version
-	Ref string `json:"ref,omitempty"`
-
-	// Type specifies the type of installation specification
-	// that could be provided as part of a custom resource.
-	// This time is used in codec to successfully decode from raw extensions.
-	// +kubebuilder:validation:Enum=helm-chart;oci-ref;"kustomize";""
-	Type RefTypeMetadata `json:"type,omitempty"`
-
-	// CredSecretSelector is an optional field, for OCI image saved in private registry,
-	// use it to indicate the secret which contains registry credentials,
-	// must exist in the namespace same as manifest
-	CredSecretSelector *metav1.LabelSelector `json:"credSecretSelector,omitempty"`
-}
-
-type RefTypeMetadata string
-
-func (t RefTypeMetadata) NotEmpty() bool {
-	return t != NilRefType
-}
-
-const (
-	OciRefType RefTypeMetadata = "oci-ref"
-	NilRefType RefTypeMetadata = ""
-)
-
-func GetSpecType(data []byte) (RefTypeMetadata, error) {
-	raw := make(map[string]json.RawMessage)
-	if err := yaml.Unmarshal(data, &raw); err != nil {
-		return "", fmt.Errorf("failed to get spec type for v1beta2 version: %w", err)
-	}
-
-	var refType RefTypeMetadata
-	if err := yaml.Unmarshal(raw["type"], &refType); err != nil {
-		return "", fmt.Errorf("failed to get spec type: %w", err)
-	}
-
-	return refType, nil
-}
