+++ api/v1beta2/manifest_types.go
@@ -60,6 +60,36 @@ type ManifestSpec struct {
 	Resource *unstructured.Unstructured `json:"resource,omitempty"`
 }
 
+// ImageSpec defines OCI Image specifications.
+// +k8s:deepcopy-gen=true
+type ImageSpec struct {
+	// Repo defines the Image repo
+	Repo string `json:"repo,omitempty"`
+
+	// Name defines the Image name
+	Name string `json:"name,omitempty"`
+
+	// Ref is either a sha value, tag or version
+	Ref string `json:"ref,omitempty"`
+
+	// Type specifies the type of installation specification
+	// that could be provided as part of a custom resource.
+	// This time is used in codec to successfully decode from raw extensions.
+	// +kubebuilder:validation:Enum=helm-chart;oci-ref;"kustomize";""
+	Type RefTypeMetadata `json:"type,omitempty"`
+
+	// CredSecretSelector is an optional field, for OCI image saved in private registry,
+	// use it to indicate the secret which contains registry credentials,
+	// must exist in the namespace same as manifest
+	CredSecretSelector *metav1.LabelSelector `json:"credSecretSelector,omitempty"`
+}
+
+type RefTypeMetadata string
+
+const (
+	OciRefType RefTypeMetadata = "oci-ref"
+)
+
 // ManifestStatus defines the observed state of Manifest.
 type ManifestStatus declarative.Status
